Staged changes that have been left waiting too long.
This probably has something to do with needing to make a bunch of
changes to binrw followed by also making similar changes to binrw_derive
and then i never made the changes to binrw_derive.

of course now binrw has been redone so this is basically just put up for
archival purposes.